refactor: simplify packages/errors imports
