+++ src/clients/api/mutations/swapTokens/index.ts
@@ -1,9 +1,8 @@
 import { ContractTransaction } from 'ethers';
 import { SwapRouter } from 'packages/contracts';
+import { VError } from 'packages/errors';
 import { Swap } from 'types';
 import { generateTransactionDeadline } from 'utilities';
-
-import { VError } from 'packages/errors/VError';
 
 export interface SwapTokensInput {
   swapRouterContract: SwapRouter;
