+++ src/clients/api/queries/getVoterHistory/index.ts
@@ -1,7 +1,6 @@
+import { VError } from 'packages/errors';
 import { VoterHistory } from 'types';
 import { restService } from 'utilities';
-
-import { VError } from 'packages/errors/VError';
 
 import formatVoterHistoryResponse from './formatVoterHistoryResponse';
 import { GetVoterHistoryResponse } from './types';
