+++ src/hooks/useOperationModal/Modal/SupplyForm/index.tsx
@@ -11,7 +11,7 @@ import {
   TokenTextField,
 } from 'components';
 import { useGetSwapRouterContractAddress } from 'packages/contracts';
-import { displayMutationError } from 'packages/errors';
+import { VError, displayMutationError } from 'packages/errors';
 import { isTokenActionEnabled } from 'packages/tokens';
 import { useTranslation } from 'packages/translations';
 import React, { useCallback, useMemo, useState } from 'react';
@@ -27,7 +27,6 @@ import useGetSwapInfo from 'hooks/useGetSwapInfo';
 import useGetSwapTokenUserBalances from 'hooks/useGetSwapTokenUserBalances';
 import { useIsFeatureEnabled } from 'hooks/useIsFeatureEnabled';
 import useTokenApproval from 'hooks/useTokenApproval';
-import { VError } from 'packages/errors/VError';
 
 import { useStyles as useSharedStyles } from '../styles';
 import Notice from './Notice';
