+++ src/packages/errors/displayMutationError/__tests__/index.spec.ts
@@ -1,7 +1,6 @@
+import { VError } from 'packages/errors';
 import { displayNotification } from 'packages/notifications';
 import Vi from 'vitest';
-
-import { VError } from 'packages/errors/VError';
 
 import { displayMutationError } from '..';
