+++ src/packages/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './transactionErrors';
 export * from './displayMutationError';
 export * from './logError';
+export * from './VError';
