+++ src/packages/errors/transactionErrors.ts
@@ -1,8 +1,7 @@
 import BigNumber from 'bignumber.js';
 import { ContractReceipt } from 'ethers';
 
-import { VError, VErrorPhraseMap } from 'packages/errors/VError';
-
+import { VError, VErrorPhraseMap } from './VError';
 import {
   ComptrollerErrorReporterError,
   ComptrollerErrorReporterFailureInfo,
