+++ src/pages/Governance/ProposalList/CreateProposalModal/checkImportErrors/index.ts
@@ -1,6 +1,5 @@
 import { FormikErrors } from 'formik';
-
-import { VError } from 'packages/errors/VError';
+import { VError } from 'packages/errors';
 
 import { FormValues } from '../proposalSchema';
